+++ packages/shared/CHANGELOG.md
@@ -1,5 +1,57 @@
 # Change Log
 
+## 3.6.0
+
+### Minor Changes
+
+- Export `createEventBus` from `@clerk/shared/eventBus`. ([#5546](https://github.com/clerk/javascript/pull/5546)) by [@panteliselef](https://github.com/panteliselef)
+
+  ```ts
+  // Create a type-safe event bus
+  const bus = createEventBus<{
+    'user:login': { id: string };
+    error: Error;
+  }>();
+
+  // Subscribe to events
+  const onLogin = ({ id }: { id: string }) => console.log('User logged in:', id);
+  bus.on('user:login', onLogin);
+
+  // Subscribe with priority (runs before regular handlers)
+  bus.onBefore('error', error => console.error('Error occurred:', error));
+
+  // Emit events
+  bus.emit('user:login', { id: 'user_123' });
+
+  // Unsubscribe specific handler
+  bus.off('user:login', onLogin);
+
+  // Unsubscribe all handlers
+  bus.off('error');
+  ```
+
+### Patch Changes
+
+- Improve JSDoc comments ([#5575](https://github.com/clerk/javascript/pull/5575)) by [@LekoArts](https://github.com/LekoArts)
+
+- Fix JWT v2 feature parsing ([#5580](https://github.com/clerk/javascript/pull/5580)) by [@octoper](https://github.com/octoper)
+
+- Updated dependencies [[`554242e`](https://github.com/clerk/javascript/commit/554242e16e50c92a6afb6ed74c681b04b9f113b5)]:
+  - @clerk/types@4.53.0
+
+## 3.5.0
+
+### Minor Changes
+
+- Adding the new `o` claim that contains all organization related info for JWT v2 schema ([#5549](https://github.com/clerk/javascript/pull/5549)) by [@octoper](https://github.com/octoper)
+
+### Patch Changes
+
+- Add Payment Sources to `<OrgProfile />`, hook up all org-related payment source and checkout methods to the org-specific endpoints ([#5554](https://github.com/clerk/javascript/pull/5554)) by [@aeliox](https://github.com/aeliox)
+
+- Updated dependencies [[`3ad3bc8`](https://github.com/clerk/javascript/commit/3ad3bc8380b354b0cd952eb58eb6c07650efa0f2), [`cfa94b8`](https://github.com/clerk/javascript/commit/cfa94b88476608edf8c2486e8ec0d3f3f82e0bfb), [`2033919`](https://github.com/clerk/javascript/commit/203391964857b98dae11944799d1e6328439e838), [`5f3cc46`](https://github.com/clerk/javascript/commit/5f3cc460b6b775b5a74746758b8cff11649a877a)]:
+  - @clerk/types@4.52.0
+
 ## 3.4.1
 
 ### Patch Changes